requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy
matplotlib
seaborn
scikit-learn
